pyproject.toml
[project]
name = "pdf2notion"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "gokamoda", email = "gokamoda@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "click>=8.2.0",
    "notion-client>=2.3.0",
    "pymupdf>=1.25.5",
    "python-gyazo>=2.3.0",
    "requests>=2.32.3",
    "tqdm>=4.67.1",
]


[project.urls]
Repository = "https://github.com/gokamoda/pdf2notion"

[project.scripts]
pdf2notion = "pdf2notion:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = "src"
testpaths = ["src",]

[tool.ruff]
exclude = [".git", ".mypy_cache", ".ruff_cache", ".venv", "third_party"]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
fixable = ["ALL"]
unfixable = []
select = [
    "A",  # flake8-builtin
    "B",  # flake8-bugbear
    "E",  # pycodestyle error
    "F",  # Pyflakes
    "I",  # isort
    "N",  # pep8-naming
    "W",  # pycodestyle warning
    "PL",  # Pylint
    "UP",  # pyupgrade
]
ignore = [
    "B905",  # Zip-without-explicit-strict
    "E501",  # Line too long, handled by black
    "F401",  # module imported but unused
    "F403",  # Unable to detect undefined names
    "N812",  # Lowercase imported as non-lowercase
    "N999",  # Invalid module name
    "UP035",  # Checks for uses of deprecated imports based on the minimum supported Python version.
    "PLR0912",  # Too many branches
    "PLR0913",  # Too many arguments in function definition
    "PLR0915",  # Too many statements in function
    "PLR2004",  # Magic value used in comparison
]

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true  # Settings for Python 3.8 compatibility

[tool.ruff.format]
quote-style = "double"
line-ending = "auto"

[tool.mypy]
show_error_context = true
show_column_numbers = true
ignore_missing_imports = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
warn_unused_ignores = true
warn_redundant_casts = true
